VOICE_RECORDER/requirements.txt
Pillow==12.0.0
scipy==1.16.3
sounddevice==0.5.3
wavio==0.0.9
